Warn about problems with libssl and libcrypto packaging if Python needs them.
Found an annoying problem with compiling against Python.  The
configure check would fail because it couldn't link against libssl and
libcrypto, neither which is used here.  However, Python wanted them,
but the symbolic links for these libraries were not present on
Debian-based systems even though the packages are installed.  It turns
out the dev packages for these libraries must be installed for
AC_PYTHON_DEVEL (which uses Python's distutils to determine library
dependencies) to be satisfied.

Solution is to either not use AC_PYTHON_DEVEL or to just warn the user
that these dev packages need to be installed, so I chose the latter.
Unfortunately AC_PYTHON_DEVEL (called by SWIG_PYTHON) doesn't give us
an opportunity to provide a message on failure, so we have to add the
warning just before calling it.  This doesn't seem to be a problem on
OS X, but more testing is needed.