libmapper: rename 'announce' functions to 'probe' and don't specialize in on_collision().
Since proposed resources are announced using the '/probe' message, as
differentiated from registered resources that are announced using the
'/registered' message, it is better if the source reflects this
terminology.  Now instead of a specialized switch in on_collision(),
we can use a callback to handle formatting the '/registered' message,
so that the resource-allocation logic remains independent of message
formatting.  (e.g., the name.ordinal needs special processing to
format the '/registered' message.)