libmapper: use double-pointers as iterators to walk through a database query.
The user must dereference the pointer twice to access the record, and
he passes it back in to get the next result.

Some nice properties of this approach: In the case of walking all
records, no memory allocation is necessary.  In the case of searching,
a continuation context is passed to allow lazy computation, and
reduced memory allocation compared to the eager approach since the
iterator is reused.  Both cases are transparent to the user.

A test program 'testdb' is included.