Make it possible to allocate an admin and pass it to a device or monitor.
This changes the mdev_new() interface, though if 'iface' parameter is
zero there is no problem.  Now in rare cases where iface, group, or
port should be customized, the user can create a custom admin and pass
it to device or monitor.  Otherwise if it is 0, device or monitor
creates and admin with standard values.