Add a little string-based lookup table implementation using qsort and bsearch.
This uses a sorted array of key-value pairs to store pointers.
Insertion should be considered as "slow" (around O(n log n)), look-up
as "fast" (around O(log n)).  We could always replace it with a hash
table in the future if we want O(1), but this should be good enough
for now and is a quick and easy solution.