Fix issue with list_header_t on Darwin 64-bit.
Seems the zero-length array used as a stub in list_header_t caused
problematic behaviour on Snow Leopard 64-bit, so use a one-value array
and calculate the size explicitly which should be system independant.
Had to rearrage the structure a little too, since the packing of the
enum between void pointers caused extra space to be padded in the case
of 64-bit architecture.