Add a simple PWM synthesizer to use as a sound example.
This patch pulls in RtAudio 4.0.7 into the project as a minimal way to
access platform sound APIs.  configure.ac was modified to detect ALSA
or CoreAudio and provide the RtAudio compiler flags accordingly.

The file `pwm_example` runs this synth with mappable parameters as a
command-line tool.

In order to provide a GUI, it also includes a second Python project
`py_tk_gui` which wraps the synth in a Python module and uses Tkinter
to provide mappable knobs.  This will of course only work if SWIG is
found and enabled.

Note that because of RtAudio's C++iness, the examples provided are
.cpp files to make autoconf figure out that it needs to use g++ for
linking. It might be nice in the future to come up with a C++less
alternative, but this'll have to do for now.