Remove type-specific "shortcuts" for adding signals, use void* instead.
Realized it would be a mistake to suggest that you can pass an array
of mapper_signal_value_t to this function, since the memory layout
might not remain the same when we add new types, and we can't count on
it lining up with float or int arrays.  Instead, just use void*.  This
makes C perform pointer coercion without a warning, so the need for a
macro to convert the pointer to mapper_signal_value_t* goes away.

With the observation that calling the full mdev_add_input/output is
actually now less characters that using mdev_add_float_input et al.,
may as well remove these as they aren't really needed or helpful.