Add a check to disable vector mapping.
Unfortunately the expression compiler and evaluator both need some
work before vector mappings are handled properly.

Better to disable it for now, since it won't be ready for first
release--considered interesting future work:

Basically, currently an `x' in an expression is considered the same as
`x[0]'.  This is wrong, since we don't support a way to build vectors,
and therefore it implies that only vector-to-scalar mappings are
possible.

Two things need to happen: we need a `comma' (e.g. Python) or `space'
(e.g. Matlab) operator to allow building vectors in an expression:

y = [ x[0]*x[1], x[1]*x[2], x[0]*x[2] ]

We also need to treat a long `x' as a reference to the entire vector,
treated as a first-class value.  The evaluator currently assumes the
stack contains only scalars.  We need to be able to push an entire
vector on the stack and perform operations across its contents.  That
is, the expression,

y = x*10

should multiple each item in `x' by 10, whatever its size.