Proposal for alternative instances API.
Uses same implementation internally, but instead of exposing a
user-facing mapper_signal_instance data type, it allows user to
specify an external ID, and attach an optional data pointer.

Allows for consistency in instance IDs across several signals, so that
user software can interpret this as representative of an "object"
formed of a collection of signal instances.