Reestablish roles between signal/device/router for updating instance values.
This patch may seem a bit arbitrary, so some explanation is warranted.
For example, it moves `msig_send_instance` into device.c, naming it
`mdev_route_instance`, in which the passed-in device pointer is not
even used.

Previously, we had a structure where signals, when updated, requested
that the device send out their new values to the routers:

           Device
          /      \
       Signal   Router
       (Input)    |
                Sig Conn
                  |
                Connection
                (Output)

In the instances branch, some data per-instance, per-signal,
per-connection is required because the output history of expressions
may differ between connections for the same signal, on a per-instance
basis.  We now have something more like,

           Device
          /      \
       Signal   Router
          |         \
       Sig Inst   Sig Conn
       (Input)        \
          |            \
       Conn Inst --> Connection
       (Output)

Ideally, the signal shouldn't know anything about links or
connections, but since history information is per-signal and
per-connection, it was most convenient and efficient to point directly
to the connection-specific data from the signal.

Unfortunately, I believe that this odd loop in the data structures,
besides breaking some principles of opaqueness and separation of
concerns, will almost certainly lead to problems in the future if we
want to support the idea of remote routers.

So this patch splits up handling of per-connection data processing
into a service provided by the device, which hands off processing to
the concerned routers.  However, no look-up in the device or router
structures is needed, because the signal already has this information.

Therefore it may seem arbitrary since the device and router are not
really needed except for the final transport layer, however I think
`mapper_connection_instance` should eventually be made opaque to code
in `signal.c`, where signal transformation and handling of
`mapper_connection_instance` should be performed entirely in
`router.c`.

Mainly this change is to help in performing the merge with more recent
code in other branches, since it is difficult to understand the merge
conflicts when core functions, also modified by other branches, have
been eliminated and concentrated in `signal.c`.