Handle bundle mechanics implicitly in the router.
Makes sense for device to just tell router to simply receive new
signal values, and when to dispatch received messages together.
Therefore instead of passing it a bundle explicitly, router builds up
the bundle as necessary if it hasn't yet received a "send"
instruction. This is of course more stateful and less friendly to
concurrency, but, I think, cleaner, as the device doesn't need to
worry about allocating resources at the transport layer.