Working on a major reorganization of the router structure, in order to restore modularity of devices and routers with the new instances functionality. Since I am already rewriting large parts of the router I decided to try to reconcile the timetag/queue functionality with the rate/count changes, and to finish adding timetags to the API as needed. In future commits I will finish working out the rate stuff, this commit focuses on moving signal processing memory from the signal data structure to the router. The routers' conception of a mapped signal has been renamed from mapper_signal_connection to mapper_router_signal; as before this structure contains the connections for a given signal, but now it also contains the signal instance input histories if required by the defined expression.
In order to handle passing vectors of signals to the msig_update() functions and still allow queueing, it will be necessary to build the message bundles in the routers as we go rather than queueing a list of updated signal instances, since we need to pass a pointer to the vector to the device/routers. For this reason, I have modified the timetagging/queueing API such that the user simply passes a timetag with each update/release/query message; this timetag will be passed with the message when sent and the message will be sent immediately unless the use first calls mdev_start_queue() with the same timetag. Queues are now indexed only by timetag, and an arbitrary number of queues can be built simultaneously.