Implement a custom transport mediated by signal/connection properties.
On connection, this program establishes a TCP connection during the
connection callback and uses it instead of msig_update().

Therefore, in this case libmapper is used only to determine the
destination IP and to establish the connection, but it performs no
data transformation or transmission.  Rather, the program specifies
this itself, determined by specifying a custom transport in the signal
properties which are copied to the /connectTo message.