hander_map_to is a graph handler
it doesn't do anything with the device, it just immediately looks
for a map to create or update. Every local device does not need to
do this seperately since they will all find the same map over and
over and only the first one to handle the callback will actually take
any meaningful action.