changed handler_name and handler_name_probe into pseudo graph handlers
these are properly device-oriented handlers, but if multiple callbacks
with the same address pattern are registered in liblo, only the first
one registered ever gets called (radarsat1/liblo#116). I've worked
around this by having these register to the graph, and then having the
graph pass the callbacks on to each local device.

I also fixed an operator precedence bug that was causing segfaults
(++dev... -> dev... += 1;), and added an explicit cast to testprops that
the compiler was complaining about while I was debugging using that
test.

I also also added a simple test program to verify that two devices can
coexist with a shared graph and still get ready in a reasonable amount
of time. I would appreciate if someone would help integrate it into the
main test suite that is run by `make tests` from the test directory.